Fix(db): correct uploader_rating type to float in migration and seeds

+++ lib/le_meilleur_banc/listings/bench.ex
@@ -10,7 +10,7 @@ defmodule LeMeilleurBanc.Listings.Bench do
     field :longitude, :float
     field :photo_url, :string
     field :uploader_comment, :string
-    field :uploader_rating, :integer
+    field :uploader_rating, :float
     field :user_id, :id
 
     timestamps(type: :utc_datetime)
@@ -19,7 +19,25 @@ defmodule LeMeilleurBanc.Listings.Bench do
   @doc false
   def changeset(bench, attrs) do
     bench
-    |> cast(attrs, [:name, :address, :latitude, :longitude, :photo_url, :uploader_comment, :uploader_rating, :status])
-    |> validate_required([:name, :address, :latitude, :longitude, :photo_url, :uploader_comment, :uploader_rating, :status])
+    |> cast(attrs, [
+      :name,
+      :address,
+      :latitude,
+      :longitude,
+      :photo_url,
+      :uploader_comment,
+      :uploader_rating,
+      :status
+    ])
+    |> validate_required([
+      :name,
+      :address,
+      :latitude,
+      :longitude,
+      :photo_url,
+      :uploader_comment,
+      :uploader_rating,
+      :status
+    ])
   end
 end
